+++ src/common/context.c
@@ -10,7 +10,7 @@
 
 #define DEFAULT_INPUT_FILE "sipnet.in"
 #define DEFAULT_FILE_NAME "sipnet"
-#define DEFAULT_EVENTS_FILE "events"
+#define DEFAULT_EVENTS_PREFIX "events"
 #define NO_DEFAULT_FILE ""
 #define ARG_OFF 0
 #define ARG_ON 1
@@ -26,6 +26,8 @@ static char keyName[CONTEXT_CHAR_MAXLEN];
 void initContext(void) {
   // Init hash map to NULL before adding anything to it
   ctx.metaMap = NULL;
+  ctx.eventsInFile[0] = '\0';
+  ctx.eventsOutFile[0] = '\0';
 
   // clang-format off
   // Init the params
@@ -53,7 +55,7 @@ void initContext(void) {
   CREATE_CHAR_CONTEXT(climFile,       "CLIM_FILE",        NO_DEFAULT_FILE);
   CREATE_CHAR_CONTEXT(outFile,        "OUT_FILE",         NO_DEFAULT_FILE);
   CREATE_CHAR_CONTEXT(outConfigFile,  "OUT_CONFIG_FILE",  NO_DEFAULT_FILE);
-  CREATE_CHAR_CONTEXT(eventsFile,     "EVENTS_FILE",      DEFAULT_EVENTS_FILE);
+  CREATE_CHAR_CONTEXT(eventsPrefix,   "EVENTS_PREFIX",    DEFAULT_EVENTS_PREFIX);
   CREATE_CHAR_CONTEXT(inputFile,      "INPUT_FILE",       DEFAULT_INPUT_FILE);
   CREATE_CHAR_CONTEXT(restartIn,      "RESTART_IN",       NO_DEFAULT_FILE);
   CREATE_CHAR_CONTEXT(restartOut,     "RESTART_OUT",      NO_DEFAULT_FILE);