+++ tests/sipnet/test_events_types/testEventIrrigation.c
@@ -32,7 +32,7 @@ int run(void) {
 
   //// ONE IRRIGATION EVENT
   // amount 5, method 1 (soil)
-  initEvents("events_one_irrig.in", 0);
+  initEvents("events_one_irrig.in", "events.out", 0);
   setupEvents();
   procEvents();
   // should have 5 going to the soil
@@ -41,7 +41,7 @@ int run(void) {
   //// TWO IRRIGATION EVENTS
   // amount 3, method 1 (soil)
   // amount 4, method 0 (canopy)
-  initEvents("events_two_irrig.in", 1);
+  initEvents("events_two_irrig.in", "events.out", 1);
   setupEvents();
   procEvents();
   // event 1: 3 to soil