+++ tests/sipnet/test_events_types/testEventPlanting.c
@@ -42,7 +42,7 @@ int run(void) {
   initEnv();
 
   //// ONE PLANTING EVENT
-  initEvents("events_one_planting.in", 0);
+  initEvents("events_one_planting.in", "events.out", 0);
   setupEvents();
   procEvents();
   // added: leaf 10, wood 5, fine root 4, coarse root 3
@@ -50,7 +50,7 @@ int run(void) {
 
   //// TWO PLANTING EVENTS
   initEnv();
-  initEvents("events_two_planting.in", 1);
+  initEvents("events_two_planting.in", "events.out", 1);
   setupEvents();
   procEvents();
   // leaf 10+9, wood 5+6, fine root 4+8, coarse root 3+4